Move Button to components/lib — a directory for our future components library
