+++ pages/event-decoder.page.tsx
@@ -3,7 +3,7 @@ import { addHexPrefix } from 'ethereumjs-util';
 import { ChangeEvent, useMemo, useState } from 'react';
 
 import DecoderSvg from '../public/static/svg/decoders';
-import { Button } from '../src/components/Button';
+import { Button } from '../src/components/lib/Button';
 import { Spinner } from '../src/components/Spinner';
 import { ToolLayout } from '../src/layout/ToolLayout';
 import { decodeWithEventProps } from '../src/lib/decodeBySigHash';
